Working on build system with C++ (linker error)

+++ ubuntu-asf/src/peripheral/led/led0.h
@@ -1,7 +1,7 @@
 #ifndef PERIPHERAL_LED_LED0_H_ 
 #define PERIPHERAL_LED_LED0_H_ 
 
-#include <stdint.h>
+//#include <stdint.h>
 
 #include "peripheral/led/gpio-led-interface.h"
 
@@ -17,7 +17,7 @@ class Led0 : public GpioLedInterface {
  private:
   auto SetAsOutput() const -> void;
   static bool instantiated_;
-  const uint8_t pin_;
+  const /*uint8_t*/ int pin_;
 };
 
 }  // namespace peripheral
